pyproject.toml
[project]
name = "scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "libusb1>=3.3.1",
    "pyinstaller>=6.14.1",
    "python-escpos>=3.1",
    "pyusb>=1.3.1",
]
